+++ src/RcppExports.cpp
@@ -278,6 +278,19 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// vinecop_pdf_full_cpp
+Rcpp::List vinecop_pdf_full_cpp(const Eigen::MatrixXd& u, const Rcpp::List& vinecop_r, size_t cores);
+RcppExport SEXP _rvinecopulib_vinecop_pdf_full_cpp(SEXP uSEXP, SEXP vinecop_rSEXP, SEXP coresSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type u(uSEXP);
+    Rcpp::traits::input_parameter< const Rcpp::List& >::type vinecop_r(vinecop_rSEXP);
+    Rcpp::traits::input_parameter< size_t >::type cores(coresSEXP);
+    rcpp_result_gen = Rcpp::wrap(vinecop_pdf_full_cpp(u, vinecop_r, cores));
+    return rcpp_result_gen;
+END_RCPP
+}
 // vinecop_cdf_cpp
 Eigen::VectorXd vinecop_cdf_cpp(const Eigen::MatrixXd& u, const Rcpp::List& vinecop_r, size_t N, size_t cores, std::vector<int> seeds);
 RcppExport SEXP _rvinecopulib_vinecop_cdf_cpp(SEXP uSEXP, SEXP vinecop_rSEXP, SEXP NSEXP, SEXP coresSEXP, SEXP seedsSEXP) {
@@ -290,6 +303,48 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< size_t >::type cores(coresSEXP);
     Rcpp::traits::input_parameter< std::vector<int> >::type seeds(seedsSEXP);
     rcpp_result_gen = Rcpp::wrap(vinecop_cdf_cpp(u, vinecop_r, N, cores, seeds));
+    return rcpp_result_gen;
+END_RCPP
+}
+// vinecop_scores_cpp
+Eigen::MatrixXd vinecop_scores_cpp(const Eigen::MatrixXd& u, const Rcpp::List& vinecop_r, bool step_wise, size_t cores);
+RcppExport SEXP _rvinecopulib_vinecop_scores_cpp(SEXP uSEXP, SEXP vinecop_rSEXP, SEXP step_wiseSEXP, SEXP coresSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type u(uSEXP);
+    Rcpp::traits::input_parameter< const Rcpp::List& >::type vinecop_r(vinecop_rSEXP);
+    Rcpp::traits::input_parameter< bool >::type step_wise(step_wiseSEXP);
+    Rcpp::traits::input_parameter< size_t >::type cores(coresSEXP);
+    rcpp_result_gen = Rcpp::wrap(vinecop_scores_cpp(u, vinecop_r, step_wise, cores));
+    return rcpp_result_gen;
+END_RCPP
+}
+// vinecop_hessian_cpp
+Eigen::MatrixXd vinecop_hessian_cpp(const Eigen::MatrixXd& u, const Rcpp::List& vinecop_r, bool step_wise, size_t cores);
+RcppExport SEXP _rvinecopulib_vinecop_hessian_cpp(SEXP uSEXP, SEXP vinecop_rSEXP, SEXP step_wiseSEXP, SEXP coresSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type u(uSEXP);
+    Rcpp::traits::input_parameter< const Rcpp::List& >::type vinecop_r(vinecop_rSEXP);
+    Rcpp::traits::input_parameter< bool >::type step_wise(step_wiseSEXP);
+    Rcpp::traits::input_parameter< size_t >::type cores(coresSEXP);
+    rcpp_result_gen = Rcpp::wrap(vinecop_hessian_cpp(u, vinecop_r, step_wise, cores));
+    return rcpp_result_gen;
+END_RCPP
+}
+// vinecop_scores_cov_cpp
+Eigen::MatrixXd vinecop_scores_cov_cpp(const Eigen::MatrixXd& u, const Rcpp::List& vinecop_r, bool step_wise, size_t cores);
+RcppExport SEXP _rvinecopulib_vinecop_scores_cov_cpp(SEXP uSEXP, SEXP vinecop_rSEXP, SEXP step_wiseSEXP, SEXP coresSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type u(uSEXP);
+    Rcpp::traits::input_parameter< const Rcpp::List& >::type vinecop_r(vinecop_rSEXP);
+    Rcpp::traits::input_parameter< bool >::type step_wise(step_wiseSEXP);
+    Rcpp::traits::input_parameter< size_t >::type cores(coresSEXP);
+    rcpp_result_gen = Rcpp::wrap(vinecop_scores_cov_cpp(u, vinecop_r, step_wise, cores));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -387,7 +442,11 @@ static const R_CallMethodDef CallEntries[] = {
     {"_rvinecopulib_vinecop_rosenblatt_cpp", (DL_FUNC) &_rvinecopulib_vinecop_rosenblatt_cpp, 5},
     {"_rvinecopulib_vinecop_sim_cpp", (DL_FUNC) &_rvinecopulib_vinecop_sim_cpp, 5},
     {"_rvinecopulib_vinecop_pdf_cpp", (DL_FUNC) &_rvinecopulib_vinecop_pdf_cpp, 3},
+    {"_rvinecopulib_vinecop_pdf_full_cpp", (DL_FUNC) &_rvinecopulib_vinecop_pdf_full_cpp, 3},
     {"_rvinecopulib_vinecop_cdf_cpp", (DL_FUNC) &_rvinecopulib_vinecop_cdf_cpp, 5},
+    {"_rvinecopulib_vinecop_scores_cpp", (DL_FUNC) &_rvinecopulib_vinecop_scores_cpp, 4},
+    {"_rvinecopulib_vinecop_hessian_cpp", (DL_FUNC) &_rvinecopulib_vinecop_hessian_cpp, 4},
+    {"_rvinecopulib_vinecop_scores_cov_cpp", (DL_FUNC) &_rvinecopulib_vinecop_scores_cov_cpp, 4},
     {"_rvinecopulib_vinecop_select_cpp", (DL_FUNC) &_rvinecopulib_vinecop_select_cpp, 22},
     {"_rvinecopulib_vinecop_fit_cpp", (DL_FUNC) &_rvinecopulib_vinecop_fit_cpp, 10},
     {"_rvinecopulib_fit_margins_cpp", (DL_FUNC) &_rvinecopulib_fit_margins_cpp, 9},